Tag 2020-06-17 generated automatically from AcquiaRA-2020-06-17-21-39-42

+++ docroot/modules/contrib/dx8/modules/cohesion_website_settings/src/Entity/IconLibrary.php
@@ -2,14 +2,10 @@
 
 namespace Drupal\cohesion_website_settings\Entity;
 
-use Drupal\cohesion\Entity\CohesionConfigEntityBase;
 use Drupal\cohesion\EntityHasResourceObjectTrait;
 use Drupal\Core\Entity\EntityStorageInterface;
-use Drupal\Component\Serialization\Json;
-use Drupal\Core\Cache\Cache;
 use Drupal\Core\StringTranslation\TranslatableMarkup;
 use Drupal\cohesion\Entity\CohesionSettingsInterface;
-use Drupal\cohesion_website_settings\Plugin\Api\WebsiteSettingsApi;
 
 /**
  * Defines the Cohesion website settings entity.
@@ -40,6 +36,17 @@ use Drupal\cohesion_website_settings\Plugin\Api\WebsiteSettingsApi;
  *   links = {
  *     "in-use" = "/admin/cohesion/cohesion_icon_library/{cohesion_icon_library}/in_use",
  *     "collection" = "/admin/cohesion/cohesion_website_settings"
+ *   },
+ *   config_export = {
+ *     "id",
+ *     "label",
+ *     "json_values",
+ *     "json_mapper",
+ *     "last_entity_update",
+ *     "locked",
+ *     "modified",
+ *     "selectable",
+ *     "source"
  *   }
  * )
  */
@@ -60,13 +67,13 @@ class IconLibrary extends WebsiteSettingsEntityBase implements CohesionSettingsI
   protected $label_collection = '';
 
   /**
-   * Return all the icons combined for the form[]
+   * Return all the icons combined for the form[].
    *
-   * @return array|\stdClass|string
+   * @return array|object|string
    */
   public function getResourceObject() {
-    /** @var WebsiteSettingsApi $send_to_api */
-    $send_to_api = $this->apiProcessorManager()->createInstance('website_settings_api');
+    /** @var \Drupal\cohesion_website_settings\Plugin\Api\WebsiteSettingsApi $send_to_api */
+    $send_to_api = $this->getApiPluginInstance();
 
     return $send_to_api->getIconGroup();
   }
@@ -82,20 +89,22 @@ class IconLibrary extends WebsiteSettingsEntityBase implements CohesionSettingsI
         $original_json_values = $original->getDecodedJsonValues();
         $json_values = $this->getDecodedJsonValues();
 
-        // Clear the previous font files if the new icon library files are different or as no files,
+        // Clear the previous font files if the new icon library files are different or as no files,.
         if (isset($json_values['fontFiles'])) {
           if ((isset($original_json_values['fontFiles']) && $json_values['fontFiles'] != $original_json_values['fontFiles'])) {
-            $this->clearIconFontFiles($original_json_values);
+            $this->clearIconFontFiles($original_json_values, $json_values);
           }
-        } else {
+        }
+        else {
           $this->clearIconFontFiles($original_json_values);
         }
 
         if (isset($json_values['iconJSON']['json'])) {
           if ((isset($original_json_values['iconJSON']['json']) && $json_values['iconJSON']['json'] != $original_json_values['iconJSON']['json'])) {
-            $this->clearSelectionJson($original_json_values);
+            $this->clearSelectionJson($original_json_values, $json_values);
           }
-        } else {
+        }
+        else {
           $this->clearSelectionJson($original_json_values);
         }
       }
@@ -106,8 +115,8 @@ class IconLibrary extends WebsiteSettingsEntityBase implements CohesionSettingsI
    * {@inheritdoc}
    */
   public function process() {
-    /** @var WebsiteSettingsApi $send_to_api */
-    $send_to_api = $this->apiProcessorManager()->createInstance('website_settings_api');
+    /** @var \Drupal\cohesion_website_settings\Plugin\Api\WebsiteSettingsApi $send_to_api */
+    $send_to_api = $this->getApiPluginInstance();
     $send_to_api->setEntity($this);
     $send_to_api->send();
   }
@@ -125,7 +134,8 @@ class IconLibrary extends WebsiteSettingsEntityBase implements CohesionSettingsI
     parent::postSave($storage, $update);
 
     // Send settings to API if enabled and modified!.
-    if ($this->status()) {  // && $this->isModified()) {
+    // && $this->isModified()) {.
+    if ($this->status()) {
       $this->process();
     }
 
@@ -155,9 +165,12 @@ class IconLibrary extends WebsiteSettingsEntityBase implements CohesionSettingsI
    * {@inheritdoc}
    */
   public function getInUseMessage() {
-    return ['message' => ['#markup' => t('This icon library has been tracked as in use in the places listed below. You should not delete it until you have removed its use.'),],];
+    return ['message' => ['#markup' => t('This icon library has been tracked as in use in the places listed below. You should not delete it until you have removed its use.')]];
   }
 
+  /**
+   *
+   */
   public function clearData() {
     $json_value = $this->getDecodedJsonValues();
     $this->clearIconFontFiles($json_value);
@@ -166,28 +179,51 @@ class IconLibrary extends WebsiteSettingsEntityBase implements CohesionSettingsI
   }
 
   /**
-   * Clear font files for a given json values of a icon library
+   * Clear font files for a given json values of a icon library.
    *
    * @param $json_value
+   * @param $new_json_values
    */
-  private function clearIconFontFiles($json_value) {
+  private function clearIconFontFiles($json_value, $new_json_values = NULL) {
+    /** @var \Drupal\Core\File\FileSystemInterface $file_system */
+    $file_system = \Drupal::service('file_system');
     if (isset($json_value['fontFiles']) && is_array($json_value['fontFiles'])) {
       foreach ($json_value['fontFiles'] as $fontFile) {
         if (is_array($fontFile) && isset($fontFile['uri']) && file_exists($fontFile['uri'])) {
-          \Drupal::service('cohesion.local_files_manager')->deleteFileByURI($fontFile['uri']);
+          $should_delete = TRUE;
+          // Only delete if file have not the same path
+          if($new_json_values && isset($new_json_values['fontFiles']) && is_array($new_json_values['fontFiles'])) {
+            foreach ($new_json_values['fontFiles'] as $newFontFile) {
+              $ll = $file_system->realpath($newFontFile['uri']);
+              $ss = $file_system->realpath($fontFile['uri']);
+              if($file_system->realpath($newFontFile['uri']) == $file_system->realpath($fontFile['uri'])) {
+                $should_delete = FALSE;
+              }
+            }
+          }
+
+          if($should_delete) {
+            \Drupal::service('cohesion.local_files_manager')->deleteFileByURI($fontFile['uri']);
+          }
+
         }
       }
     }
   }
 
   /**
-   * Clear selection json for a given json values of a icon library
+   * Clear selection json for a given json values of a icon library.
    *
    * @param $json_value
+   * @param $new_json_values
    */
-  private function clearSelectionJson($json_value) {
+  private function clearSelectionJson($json_value, $new_json_values = NULL) {
+    /** @var \Drupal\Core\File\FileSystemInterface $file_system */
+    $file_system = \Drupal::service('file_system');
     if (isset($json_value['iconJSON']['json'])) {
-      \Drupal::service('cohesion.local_files_manager')->deleteFileByURI($json_value['iconJSON']['json']);
+      if(!$new_json_values || !isset($new_json_values['iconJSON']['json']) || $file_system->realpath($json_value['iconJSON']['json']) != $file_system->realpath($json_value['iconJSON']['json'])) {
+        \Drupal::service('cohesion.local_files_manager')->deleteFileByURI($json_value['iconJSON']['json']);
+      }
     }
   }
 
@@ -197,4 +233,5 @@ class IconLibrary extends WebsiteSettingsEntityBase implements CohesionSettingsI
   public function isLayoutCanvas() {
     return FALSE;
   }
+
 }
